requirements.txt
python-chess
Pillow
PyAutoGUI
numpy
